make a folder for contact validation
